Description: Fix compilation warning of 730release IssueNo: https://gitee.com/openharmony/graphic_ui/issues/I5MRU3 Feature or Bugfix: Bugfix Binary Source: No Signed-off-by: suwenxiang <suwenxiang@huawei.com>

+++ frameworks/src/core/components/component.cpp
@@ -1042,7 +1042,7 @@ int32_t Component::GetAnimatorValue(char *animatorValue, const int8_t index, boo
     }
 
     int32_t convertedValue =
-        isOpacity ? (strtod(value, nullptr) * ALPHA_MAX) : static_cast<int32_t>(strtol(value, nullptr, DEC));
+        isOpacity ? (strtod(value, nullptr) * ALPHA_MAX) : (strtol(value, nullptr, DEC));
     if (TransitionImpl::IsEndWith(value, "rad")) {
         uint8_t degConversionRate = 57;
         convertedValue = convertedValue * degConversionRate;
